requirements.txt
# Core scientific computing dependencies
numpy>=1.23.0,<2.0.0
numba>=0.58.0,<0.60.0
matplotlib>=3.8.0,<4.0.0
scipy>=1.11.0,<2.0.0

# MPI support (optional but recommended for parallel processing)
mpi4py>=3.1.0,<4.0.0

# Visualization
cmasher>=1.7.0,<2.0.0

# HDF5 support for data I/O
h5py>=3.9.0,<4.0.0

# Configuration file support
PyYAML>=6.0.0,<7.0.0

# Interactive notebooks (optional)
ipywidgets>=8.0.0,<9.0.0
jupyterlab>=4.0.0,<5.0.0

# Development dependencies (install with pip install -r requirements-dev.txt)
# pytest>=7.4.0,<8.0.0
# pytest-cov>=4.1.0,<5.0.0
# black>=23.0.0,<24.0.0
# ruff>=0.1.0,<0.2.0
# mypy>=1.5.0,<2.0.0
# sphinx>=7.0.0,<8.0.0
# sphinx-rtd-theme>=1.3.0,<2.0.0